bench: repair bit-rotted append_bench + extent_bench (u32->u64 offset widening)
[perf hot-path work, groundwork] The standalone EN-append benches had bit-rotted
against the u32->u64 extent-offset widening (AppendReq.commit / ReadBytesReq.offset
are now u64). Fixed:
- append_bench: 3 commit-offset computations to u64.
- extent_bench: `let mut commit = 0u32` -> 0u64 (x2) and `offset: (block*4096) as
  u32` -> as u64 (x4).

These are the measurement harness for EN write/read hot-path optimization (both
standalone in-process over loopback — no cluster needed). Baseline (sandbox,
loopback; NOT production-representative — real numbers need perf_check on
hardware, but the relative A/B is usable for CPU-path changes):
- 4KB Append: depth=1 42.7 MB/s, depth=32 217, depth=64 327.7 MB/s (~84K ops/s)
- 4KB Read: depth=1 303, depth=64 517 MB/s
- Mixed 1w+1r: 118K ops/s total

Bench-only change (no production code touched); validated by compile + run.

+++ crates/stream/benches/append_bench.rs
@@ -63,18 +63,18 @@ fn main() {
             let req = AppendReq {
                 extent_id,
                 eversion: 1,
-                commit: i * payload.len() as u32,
+                commit: i as u64 * payload.len() as u64,
                 owner_epoch: 1,
                 payload: payload.clone(),
             };
             client.call(MSG_APPEND, req.encode()).await.unwrap();
         }
-        let warmup_commit = 500 * payload.len() as u32;
+        let warmup_commit = 500u64 * payload.len() as u64;
 
         // Sequential
         let start = Instant::now();
         for i in 0..ops {
-            let commit = warmup_commit + (i as u32) * payload.len() as u32;
+            let commit = warmup_commit + (i as u64) * payload.len() as u64;
             let req = AppendReq {
                 extent_id,
                 eversion: 1,
